refactor: sweep verified single-caller/duplicate micro-shims
A repo-wide debt audit adversarially verified six trivial,
low-risk consolidations with named symbols and exact caller
counts, dropping every candidate that didn't hold up on closer
inspection (e.g. a "single-caller" helper actually called 13
times, a "dead" store with 6 live test consumers):

- Delete ROOT_ROUTING_ARGS, a byte-identical alias of
  AGENT_RUN_GLOBAL_ARGS.
- Inline installDesktopMenu(), a 9-line single-caller wrapper.
- Extract resolveSymlinkType()/fileTypeFor() (duplicated between
  nodeFilesystem.ts and the test-kernel FakePlatform.ts, differing
  only in fs handle type) into one shared
  src/platform/defaults/fsEntryTypeBits.ts.
- Fold the src/utils/prompt/ barrel (index.ts re-exporting
  promptUtils.ts) into a single file.
- Narrow packages/desktop preload/hostBridge.ts's exports to what
  production actually uses; ElectronHostBridge.vitest.mts now
  exercises the public installElectronHostBridge surface and
  imports the shared channel constants directly from
  hostBridgeChannels.ts instead of re-exports.
- Dedup the byte-identical jsonResponse() test helper in
  DeviceCodeAuth.vitest.mts / RelayTokens.vitest.mts into
  test-kernel/cli/fixtures/fetchResponses.ts.

Shrinks config/ratchets/knip-baseline.json by the four entries
this cleans up.
