+++ config/ratchets/knip-baseline.json
@@ -2,11 +2,6 @@
   "semantics": "Per-symbol baseline of knip dead-code findings (files/exports/types/duplicate-export groups), each identified by (file, category, name) -- never by line number, so unrelated edits that shift lines never spuriously mark a baseline entry as fixed. The ratchet (scripts/check-dead-code-ratchet.mjs) fails only when the current knip run contains a finding not in this list. Removing entries here is always welcome (it means the dead code was cleaned up); a genuinely new unused file/export/type must either be deleted or, if intentionally kept, added here in the same PR. Regenerate by piping `knip --reporter json` through extractFindings() and sorting with compareFindings(); do not hand-edit into disorder.",
   "findings": [
     {
-      "file": "packages/cli/src/commands/_helpers/globalArgs.ts",
-      "category": "duplicates",
-      "name": "AGENT_RUN_GLOBAL_ARGS,ROOT_ROUTING_ARGS"
-    },
-    {
       "file": "packages/desktop/src/desktopCommandSurface.ts",
       "category": "exports",
       "name": "DESKTOP_COMMAND_IDS"
@@ -50,21 +45,6 @@
       "file": "packages/desktop/src/main/platform/secretStorageWarningDialog.ts",
       "category": "exports",
       "name": "getSecretStorageWarningParentWindow"
-    },
-    {
-      "file": "packages/desktop/src/preload/hostBridge.ts",
-      "category": "exports",
-      "name": "createElectronHostBridge"
-    },
-    {
-      "file": "packages/desktop/src/preload/hostBridge.ts",
-      "category": "exports",
-      "name": "ELECTRON_WEBVIEW_MESSAGE_CHANNEL"
-    },
-    {
-      "file": "packages/desktop/src/preload/hostBridge.ts",
-      "category": "exports",
-      "name": "ELECTRON_WEBVIEW_PUSH_CHANNEL"
     },
     {
       "file": "packages/extension/src/frontend/lean/VscodeIntegration.ts",
